pyproject.toml
[project]
name = "ai-assistent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.121.1",
    "google-adk>=1.18.0",
    "opensearch-py>=3.0.0",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44", # It helps to build a DB connection and give funtion to do the query with db
    "sse-starlette>=3.0.3",
    "uvicorn>=0.38.0",
]

# TODO:
# we'll add liquibase